requirements.txt
cycler==0.10.0
decorator==4.4.2
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5
numpy==1.20.1
pandas==1.2.3
Pillow==8.1.1
pycairo==1.20.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-igraph==0.9.0
pytz==2021.1
PyYAML==5.4.1
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
texttable==1.6.3
threadpoolctl==2.1.0
